lib: obstacles: move class Rectangle to its own file
